refactor(profile): unused components and implement pagination for health pages (#34)
* refactor: remove unused admin and health-related pages, components, and types

- Deleted AdminReports, AdminSettings, AdminUsers pages.
- Removed HealthProfile page and related components.
- Cleaned up unused types from auth, health, household, and wealth domains.
- Consolidated modal and field components for better reusability.
- Updated Accounts page to include account balance fetching.
- Adjusted tests to reflect the removal of types and components.

* feat(pagination): implement pagination for doctor visits and vital readings endpoints

* Implement pagination across various health and household pages

- Added pagination functionality to DoctorVisits, Vitals, Calendar, Goals, Inventory, and PhysicalAssets pages.
- Introduced PAGE_SIZE constant for consistent page size management.
- Updated API calls to include pagination parameters (page and page size).
- Implemented pagination controls (Previous and Next buttons) with appropriate enabling/disabling logic based on the current page.
- Enhanced tests to verify pagination behavior, including requests for specific pages and total page counts derived from total sizes.

* feat: implement statement upload service and add unit and integration testing support

* feat: implement panache repositories and resource endpoints for physical assets, goals, vital readings, doctor visits, and inventory items.

* Refactor JAX-RS resources and Panache repositories to eliminate code duplication

* refactor: remove unused admin and health-related pages, components, and types

- Deleted AdminReports, AdminSettings, AdminUsers pages.
- Removed HealthProfile page and related components.
- Cleaned up unused types from auth, health, household, and wealth domains.
- Consolidated modal and field components for better reusability.
- Updated Accounts page to include account balance fetching.
- Adjusted tests to reflect the removal of types and components.

* feat(pagination): implement pagination for doctor visits and vital readings endpoints

* Implement pagination across various health and household pages

- Added pagination functionality to DoctorVisits, Vitals, Calendar, Goals, Inventory, and PhysicalAssets pages.
- Introduced PAGE_SIZE constant for consistent page size management.
- Updated API calls to include pagination parameters (page and page size).
- Implemented pagination controls (Previous and Next buttons) with appropriate enabling/disabling logic based on the current page.
- Enhanced tests to verify pagination behavior, including requests for specific pages and total page counts derived from total sizes.

* feat: implement statement upload service and add unit and integration testing support

* feat: implement panache repositories and resource endpoints for physical assets, goals, vital readings, doctor visits, and inventory items.

* Refactor JAX-RS resources and Panache repositories to eliminate code duplication

* fix: update response fields in Vitals component and tests to match API schema
refactor: change totalSize type from int to long in List*Response classes
feat: add unit tests for AmortizationCalculator with explicit asOfDate parameter
chore: update documentation and scripts for SonarQube token handling

* Refactor health.yaml to reduce maxLength for doctor_name, hospital_name, and speciality fields to 50 characters. Update API.md to use environment variable for Sonar token. Revise CICD.md to clarify backend job matrix and update branch naming conventions. Enhance E2E_TESTING.md with additional test specs and clarify service startup order. Update REQUIREMENTS_household_domain.md to reflect the implementation status of the Household domain. Fix VARCHAR length inconsistencies in health domain and address schema nullability gaps in household domain. Remove dead code from profile domain and consolidate Flyway migrations across domains. Improve error handling tests and ensure test results files are properly formatted.

* feat(api): add profile_id parameter to wealth API endpoints

- Updated getAccount, updateAccount, updateAccountClassification, deactivateAccount, and related functions to include profile_id as a query parameter.
- Modified corresponding tests to reflect changes in API calls.
- Adjusted Setup.js to pass profileId when updating account classification.
- Enhanced DoctorVisits, Vitals, Calendar, Goals, Inventory, Profiles, VacationPlanner, and Reports pages to utilize user.admin_id when listing profiles.
- Added error handling improvements in errorHandler.js and comprehensive tests for error handling utilities.

* docs(wealth): confirm profile-scoped Tier B tests pass against reset DB

Updates the wealth domain-state doc to reflect that the new
AccountPanacheRepositoryTest/PhysicalAssetPanacheRepositoryTest
cross-profile findById tests were executed (not just written) after
the v0.5.1 db-reset, and all pass.


---------

+++ application/domain/household/adapters/src/main/java/com/suchika/household/adapters/http/dto/ListCalendarEventsResponse.java
@@ -12,7 +12,7 @@ public class ListCalendarEventsResponse {
     public List<CalendarEventDto> calendarEvents;
 
     @JsonProperty("total_size")
-    public int totalSize;
+    public long totalSize;
 
     @JsonProperty("page")
     public Integer page;
@@ -31,7 +31,7 @@ public class ListCalendarEventsResponse {
      */
     public ListCalendarEventsResponse(List<CalendarEventDto> calendarEvents, long totalSize, int page, int size) {
         this.calendarEvents = calendarEvents;
-        this.totalSize = (int) totalSize;
+        this.totalSize = totalSize;
         this.page = page;
         this.size = size;
     }
